+++ internal/logic/subscription/service/subscription_timeline_list.go
@@ -14,7 +14,7 @@ import (
 
 type SubscriptionTimeLineListInternalReq struct {
 	MerchantId uint64 `json:"merchantId" dc:"MerchantId" v:"required"`
-	UserId     int    `json:"userId" dc:"Filter UserId, Default All " `
+	UserId     uint64 `json:"userId" dc:"Filter UserId, Default All " `
 	SortField  string `json:"sortField" dc:"Sort Field，gmt_create|gmt_modify" `
 	SortType   string `json:"sortType" dc:"Sort Type，asc|desc" `
 	Page       int    `json:"page"  dc:"Page, Start WIth 0" `
